+++ AdjustablePortals/Package/CHANGELOG.md
@@ -1,3 +1,11 @@
+**0.2.0**
+---
+```
+- Caching improvements for portal requirements
+- Adds portal fuel options (configurable)
+- Makes nearby piece requirements configurably optional
+```
+
 **0.1.0**
 ---
 ```